feat(activerecord): add PostgreSQL supports_* predicates, move adapter to canonical path
Move PostgreSQLAdapter from adapters/postgresql-adapter.ts to
connection-adapters/postgresql-adapter.ts (where api:compare expects
it, matching Rails' file layout). Old path re-exports for compat.

Add all 38 supports_* boolean predicates and indexAlgorithms,
matching Rails' PostgreSQLAdapter. Version-dependent methods
(supportsPartitionedIndexes, supportsVirtualColumns, etc.) query
the server via getDatabaseVersion(). supportsOptimizerHints checks
extension_available?("pg_hint_plan") matching Rails.

api:compare: postgresql_adapter.rb 0% → 71% (48/68)
